+++ mapbox_maps_flutter_web/example/patrol_test/gestures_controller_test.dart
@@ -69,7 +69,11 @@ void main() {
         tester,
         (onCreated) => MaterialApp(
           home: Scaffold(
-            body: MapWebWidget(viewport: _viewport, onMapCreated: onCreated),
+            body: MapWebWidget(
+              styleUri: 'mapbox://styles/mapbox/standard',
+              viewport: _viewport,
+              onMapCreated: onCreated,
+            ),
           ),
         ),
       );
@@ -115,7 +119,11 @@ void main() {
         tester,
         (onCreated) => MaterialApp(
           home: Scaffold(
-            body: MapWebWidget(viewport: _viewport, onMapCreated: onCreated),
+            body: MapWebWidget(
+              styleUri: 'mapbox://styles/mapbox/standard',
+              viewport: _viewport,
+              onMapCreated: onCreated,
+            ),
           ),
         ),
       );
@@ -159,7 +167,11 @@ void main() {
         tester,
         (onCreated) => MaterialApp(
           home: Scaffold(
-            body: MapWebWidget(viewport: _viewport, onMapCreated: onCreated),
+            body: MapWebWidget(
+              styleUri: 'mapbox://styles/mapbox/standard',
+              viewport: _viewport,
+              onMapCreated: onCreated,
+            ),
           ),
         ),
       );
@@ -224,7 +236,11 @@ void main() {
         tester,
         (onCreated) => MaterialApp(
           home: Scaffold(
-            body: MapWebWidget(viewport: _viewport, onMapCreated: onCreated),
+            body: MapWebWidget(
+              styleUri: 'mapbox://styles/mapbox/standard',
+              viewport: _viewport,
+              onMapCreated: onCreated,
+            ),
           ),
         ),
       );